Move <NavLink /> to it's own component
